enable auth when developing localhost
the problem is that backend and frontend run on different ports and
the browser does not accept cookies in this setup.
The solution: put backend and frontend behind a proxy
Every request to /api/** will be routed to the backend
Everything else will be routed to the frontend

The proxy is started together with the database in docker-compose
You should not need to do anything.

To connect to the proxy and use it, connect to
http://localhost:8089/