Make TextInput controlled by using a state
Fixed the bug related to TextInput not being controlled when it had to.

When setting a value into the TextInput it needs to be properly passed
using the state's setter.
When the value changes, it also needs to be updated with the state's
setter.

This resolves the warning that appeared in the console.
Might fix more problems related to async operations.